requirements.txt
numpy==1.17.3
pandas==0.25.2
bottle==0.12.17
python-dotenv==0.10.3
scikit-learn==0.22
pymongo==3.9.0
joblib==0.14.0
json5==0.8.5
dnspython==1.16.0